autoplay and last stepper tests
